Move TimeoutConn to seperate pkg dir.
Fixes #10965
